+++ wsidicom/instance/instance.py
@@ -22,10 +22,10 @@ from upath import UPath
 
 from wsidicom.errors import WsiDicomError, WsiDicomUidDuplicateError
 from wsidicom.geometry import Size, SizeMm
-from wsidicom.instance.dataset import ImageType, WsiDataset
+from wsidicom.instance.dataset import WsiDataset
 from wsidicom.instance.image_data import ImageData
 from wsidicom.instance.pillow_image_data import PillowImageData
-from wsidicom.metadata.image import ImageCoordinateSystem
+from wsidicom.metadata.image import ImageCoordinateSystem, ImageType
 from wsidicom.uid import SlideUids