+++ wsidicom/web/wsidicom_web_source.py
@@ -33,7 +33,8 @@ from wsidicom.codec import Codec
 from wsidicom.config import settings
 from wsidicom.errors import WsiDicomNotFoundError
 from wsidicom.graphical_annotations import AnnotationInstance
-from wsidicom.instance import ImageType, WsiDataset, WsiInstance
+from wsidicom.instance import WsiDataset, WsiInstance
+from wsidicom.metadata import ImageType
 from wsidicom.source import Source
 from wsidicom.thread import ConditionalThreadPoolExecutor
 from wsidicom.web.wsidicom_web_client import WsiDicomWebClient